+++ app/src/main/java/com/shinplest/mobiletermproject/ApplicationClass.java
@@ -5,9 +5,13 @@ import android.content.SharedPreferences;
 import android.util.Log;
 
 import com.naver.maps.geometry.LatLng;
+import com.naver.maps.geometry.Tm128;
+import com.naver.maps.geometry.Utmk;
 import com.shinplest.mobiletermproject.parsing.Geometry;
 
+import java.lang.reflect.Array;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import okhttp3.OkHttpClient;
@@ -24,7 +28,7 @@ public class ApplicationClass extends Application {
     // Retrofit 인스턴스
     public static Retrofit retrofit;
 
-    public static ArrayList<LatLng> testlatlng;
+    public static ArrayList<List<LatLng>> testlatlng = new ArrayList<>();
 
     public static Retrofit getRetrofit() {
         if (retrofit == null) {
@@ -46,15 +50,16 @@ public class ApplicationClass extends Application {
         double longtitude = 0;
         ArrayList<LatLng> latLngArrayList = new ArrayList<>();
 
-        ArrayList<ArrayList<Double>> path = geometry.getPath().get(1);
-        Log.d("MainActivity", "path size: " + path.size());
+        ArrayList<ArrayList<Double>> path = geometry.getPath().get(0);
+        Log.d("ApplicationClass", "path size: " + path.size());
 
         for (int i = 0; i < path.size(); i++) {
             latitude = path.get(i).get(0);
             longtitude = path.get(i).get(1);
-            Log.d("MainActivity", "latitude: " + latitude);
-            Log.d("MainActivity", "latitude: " + longtitude);
-            latLngArrayList.add(new LatLng(latitude,longtitude));
+            Log.d("ApplicationClass", "latitude: " + latitude);
+            Log.d("ApplicationClass", "latitude: " + longtitude);
+            Tm128 tm128 = new Tm128(latitude, longtitude);
+            latLngArrayList.add(tm128.toLatLng());
         }
         return latLngArrayList;
     }
